Default OptionClass.required to MISSING
If the `required` attribute is still MISSING when generating the JSON payload it will be defaulted to True.

This means that when determining a type it is possible to differentiate between the user passing `required` and it being the default, which is important because an implicit annotation should never override an explicit parameter.