Revert "Remove `_write_lock` in favour of the `closing` property"
This reverts commit b968f9cfb31a05c931c29a805f18176822650c61.

It turns out that the write lock also provides safety from another task initiating a WebSocket closure in the middle of sending the heartbeat.
The `closing` property does not protect against this.

It would also be detrimental if only a part of a WebSocket frame was sent and then the heartbeater sends some other frames.