Rethink `connect()` classmethod and provide additional safety
There are a number of changes in this commit. Most significantly is the API change.
You no longer need to use the `connect()` classmethod, instead you can directly use the instance as an asynchronous context manager.

The additional safety this commit adds is how OSError is now correctly handled by reconnecting and in the event that a heartbeat is about to be sent when the WebSocket is in the progress of reconnecting it will now wait for the reconnection rather than sleeping another duration.