Add utility functions to independently create commands
As seen in the docstrings, these functions allows commands and groups to be created in other files and later on be imported into the main app.

It is very unreasonable to enforce all commands to be defined in the same file. There has to be a way to separate them.