Rework cache to only return the old value
This was done to simplify event creation where, previously, a lot of care was put into not duplicating instances of models by having the cache return a tuple with the second item being the created instance.

That was all in vain though, as the default in-memory cache will be reworked to only store proxies of objects. Doing so, fixes another issue having to do with the immutability of models where for example messages could keep references to outdated member models.

It is believed, that other in-memory caches will work similarly and no implementation will benefit from the way the cache worked previously.