Rework event dispatching and its interaction with the cache
This commit is a squash of part of the event-rework branch which underwent several refactors as it was developed.
All refactors have been squashed into the final change for simplicity.

The point of reworking the dispatching system was to follow the same pattern as the rest of the library and its evolution. Most importantly, events will be written as dataclasses and use a separate classmethod for initialization. ErrorHandlerMixin has also been removed in favour of the built-in error handling into the dispatcher.

With this rework, came a lot of changes to the cache because of how tightly coupled they are. Most importantly, the cache will only return the previously stored value instead of a tuple, because the cache will be rewritten to only store proxies of objects. Doing so, fixes another issue having to do with the immutability of models where for example messages could keep references to outdated member models.

It is believed that other in-memory caches will work similarly and no implementation will benefit from the way the cache worked previously.

Event implementations from the event-rework branch will come in a different commit.